# Added branchless Math.Min calculation