@Turnerj 2021-11-30
Giving this another crack now...

**.NET Framework Baseline**

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1348 (21H1/May2021Update)
Intel Core i5-6600K CPU 3.50GHz (Skylake), 1 CPU, 4 logical and 4 physical cores
.NET SDK=6.0.100
  [Host]    : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  Framework : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT

Job=Framework  Runtime=.NET Framework 4.7.2
```

|        Method | NumberOfCharacters |            Mean |         Error |        StdDev |         Op/s | CacheMisses/Op | BranchMispredictions/Op | Code Size | Allocated |
|-------------- |------------------- |----------------:|--------------:|--------------:|-------------:|---------------:|------------------------:|----------:|----------:|
| Quickenshtein |                 10 |        190.9 ns |       0.28 ns |       0.26 ns | 5,237,715.05 |              0 |                       0 |   4,211 B |         - |
| Quickenshtein |                400 |    473,919.8 ns |     743.11 ns |     695.10 ns |     2,110.06 |            235 |                  70,856 |   4,211 B |         - |
| Quickenshtein |               8000 | 73,353,386.7 ns | 872,388.31 ns | 816,032.56 ns |        13.63 |         39,673 |               4,212,053 |   4,211 B |         - |