@Turnerj 2021-11-30
**With Branchless Min**

|        Method | NumberOfCharacters |            Mean |         Error |        StdDev |         Op/s | CacheMisses/Op | BranchMispredictions/Op | Code Size | Allocated |
|-------------- |------------------- |----------------:|--------------:|--------------:|-------------:|---------------:|------------------------:|----------:|----------:|
| Quickenshtein |                 10 |        219.2 ns |       0.23 ns |       0.20 ns | 4,562,704.02 |              0 |                       0 |   6,007 B |         - |
| Quickenshtein |                400 |    159,471.9 ns |   2,843.29 ns |   2,659.62 ns |     6,270.70 |             72 |                   1,251 |   6,007 B |         - |
| Quickenshtein |               8000 | 62,462,987.4 ns | 279,845.42 ns | 261,767.58 ns |        16.01 |         31,858 |                 202,191 |   6,007 B |         - |