@Turnerj 2021-11-30
❌ **Branchless Min but without Outer Loop Unrolling**

|        Method | NumberOfCharacters |            Mean |         Error |        StdDev |         Op/s | BranchMispredictions/Op | CacheMisses/Op | Code Size | Allocated |
|-------------- |------------------- |----------------:|--------------:|--------------:|-------------:|------------------------:|---------------:|----------:|----------:|
| Quickenshtein |                 10 |        212.6 ns |       0.28 ns |       0.26 ns | 4,704,467.28 |                       0 |              0 |   2,432 B |         - |
| Quickenshtein |                400 |    178,371.5 ns |     216.77 ns |     202.77 ns |     5,606.28 |                     717 |             64 |   2,432 B |         - |
| Quickenshtein |               8000 | 70,299,671.7 ns | 119,200.72 ns | 111,500.43 ns |        14.22 |                  29,559 |         29,559 |   2,432 B |         - |