@Turnerj 2021-12-02
❌ **Using SIMD in .NET Framework via Vectors**

In defense of the code - it doesn't have `ShiftRightLogical128BitLane` so my hacked implementation isn't likely optimal.

|        Method | NumberOfCharacters |             Mean |         Error |        StdDev |          Op/s | CacheMisses/Op | BranchMispredictions/Op | Code Size | Allocated |
|-------------- |------------------- |-----------------:|--------------:|--------------:|--------------:|---------------:|------------------------:|----------:|----------:|
| Quickenshtein |                 10 |         591.7 ns |       0.63 ns |       0.59 ns | 1,689,997.326 |              0 |                       0 |   2,846 B |         - |
| Quickenshtein |                400 |     745,208.6 ns |   1,060.84 ns |     992.31 ns |     1,341.906 |            309 |                     800 |   2,846 B |         - |
| Quickenshtein |               8000 | 288,599,473.3 ns | 380,012.54 ns | 355,463.96 ns |         3.465 |        132,710 |                  60,894 |   2,846 B |         - |