@Turnerj 2021-12-02
I tested the branchless code again for huge text and there really isn't any performance benefit which I find kinda strange.
While there are a lot less branch mispredictions, the time taken to process has stayed the same.

**Before**

|        Method | NumberOfCharacters |     Mean |    Error |   StdDev |  Op/s | Code Size | CacheMisses/Op | BranchMispredictions/Op | Allocated |
|-------------- |------------------- |---------:|---------:|---------:|------:|----------:|---------------:|------------------------:|----------:|
| Quickenshtein |              16000 | 252.2 ms |  2.83 ms |  2.65 ms | 3.965 |   4,211 B |        196,881 |               8,602,283 |         - |
| Quickenshtein |              32000 | 934.9 ms | 11.87 ms | 11.10 ms | 1.070 |   4,211 B |        709,427 |              19,651,243 |         - |

**After**

|        Method | NumberOfCharacters |     Mean |   Error |  StdDev |  Op/s | Code Size | CacheMisses/Op | BranchMispredictions/Op | Allocated |
|-------------- |------------------- |---------:|--------:|--------:|------:|----------:|---------------:|------------------------:|----------:|
| Quickenshtein |              16000 | 248.5 ms | 0.34 ms | 0.32 ms | 4.025 |   6,007 B |        177,948 |                 484,147 |         - |
| Quickenshtein |              32000 | 991.4 ms | 1.49 ms | 1.39 ms | 1.009 |   6,007 B |        776,602 |               1,241,361 |         - |