@Turnerj 2021-12-02
For my own curiousity, I wanted to see the scale of how the performance degraded for larger strings:

| Method | NumberOfCharacters |       Mean |     Error |    StdDev |    Op/s | Code Size | BranchMispredictions/Op | CacheMisses/Op | Allocated |
|------- |------------------- |-----------:|----------:|----------:|--------:|----------:|------------------------:|---------------:|----------:|
| Normal |               2000 |   7.459 ms | 0.0457 ms | 0.0428 ms | 134.073 |   4,211 B |                 846,460 |          3,979 |         - |
| Normal |               4000 |  22.967 ms | 0.1766 ms | 0.1652 ms |  43.541 |   4,211 B |               1,983,300 |         12,143 |         - |
| Normal |               8000 |  72.782 ms | 0.9586 ms | 0.8967 ms |  13.740 |   4,211 B |               4,094,440 |         43,769 |         - |
| Normal |              12000 | 149.794 ms | 2.5139 ms | 2.2285 ms |   6.676 |   4,211 B |               6,540,698 |         92,092 |         - |
| Normal |              16000 | 251.899 ms | 3.3323 ms | 3.1170 ms |   3.970 |   4,211 B |               8,565,009 |        174,399 |         - |


| Method | NumberOfCharacters |       Mean |     Error |    StdDev |    Op/s | Code Size | BranchMispredictions/Op | CacheMisses/Op | Allocated |
|------- |------------------- |-----------:|----------:|----------:|--------:|----------:|------------------------:|---------------:|----------:|
|     PR |               2000 |   4.018 ms | 0.0777 ms | 0.0983 ms | 248.874 |   6,007 B |                  39,919 |          1,970 |         - |
|     PR |               4000 |  15.846 ms | 0.2252 ms | 0.2106 ms |  63.108 |   6,007 B |                 103,287 |          7,415 |         - |
|     PR |               8000 |  62.540 ms | 0.3767 ms | 0.3524 ms |  15.990 |   6,007 B |                 227,908 |         32,119 |         - |
|     PR |              12000 | 140.117 ms | 0.4266 ms | 0.3782 ms |   7.137 |   6,007 B |                 362,223 |         80,896 |         - |
|     PR |              16000 | 248.454 ms | 0.4744 ms | 0.4438 ms |   4.025 |   6,007 B |                 493,522 |        135,441 |         - |

This translates to the PR being...
2000 = 1.86x faster
4000 = 1.45x faster
8000 = 1.16x faster
12000 = 1.07x faster
16000 = 1.01x faster

Not exactly ideal!